+++ blockchain/app/upgrades_test.go
@@ -80,4 +80,5 @@ func TestUpgradeHandlersRegistered(t *testing.T) {
 	// Check that upgrade handlers are registered (sampling a few key ones)
 	require.True(t, app.UpgradeKeeper.HasHandler("v1.10.0-remove-ibc"), "v1.10.0-remove-ibc upgrade handler should be registered")
 	require.True(t, app.UpgradeKeeper.HasHandler("v1.13.0"), "v1.13.0 upgrade handler should be registered")
+	require.True(t, app.UpgradeKeeper.HasHandler("v1.14.0"), "v1.14.0 upgrade handler should be registered")
 }